Finish port

+++ ExampleBot/Utils.cs
@@ -1,10 +1,10 @@
 ﻿using System;
-
+using NetCord;
 
 namespace ExampleBot;
 
 public static class Utils
 {
     public static Color GetRandomColor()
-        => new(Random.Shared.Next(0, 256), Random.Shared.Next(0, 256), Random.Shared.Next(0, 256));
+        => new((byte)Random.Shared.Next(0, 256), (byte)Random.Shared.Next(0, 256), (byte)Random.Shared.Next(0, 256));
 }
